reid-service-standalone/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
numpy
onnxruntime
Pillow
